+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 - Client scope policy [RCIAM-1241](https://jira.argo.grnet.gr/browse/RCIAM-1241)
 - Sort Identity Providers by guiOrder and displayname
 - Add authnAuthority and voPersonID user attribute to event details
+- Login events type for add, remove, suspend user from a group [RCIAM-1292](https://jira.argo.grnet.gr/browse/RCIAM-1292)
 
 ### Changed
 - Increase User Attribute Value length to 4000 [EOSC-KC-132](https://github.com/eosc-kc/keycloak/issues/132)